requirements.txt
SQLAlchemy==2.0.36
fastapi==0.115.4
python-dotenv==1.0.1
uvicorn==0.32.0
psycopg2==2.9.10
python-jose==3.3.0